test: Rename PDRE to Host in Julia variables and files
