+++ test/fixtures/hostapi-legacy/HostApiLegacyInputs.jl
@@ -1,0 +1,304 @@
+module HostApiLegacyInputs
+		# Contains (mostly valid) prefixes of the corresponding keys
+    const prefix_key_value_key_value = [
+      [
+        "stat", # prefix
+        "static", # key1
+        "Inverse", # value1
+        "even-keeled", # key2
+        "Future-proofed" # value2
+      ],
+      [
+        "Face",
+        "function",
+        "Horizontal",
+        "Face to face",
+        "Expanded"
+      ],
+      [
+        "Integrated",
+        "Integrated",
+        "portal",
+        "budgetary management",
+        "pricing structure"
+      ],
+      [
+        "non",
+        "non-based",
+        "Monitored",
+        "non-volatile",
+        "emulation"
+      ],
+      [
+        "sec",
+        "productivity",
+        "secondary",
+        "Total",
+        "Visionary"
+      ],
+      [
+        "conc",
+        "Exclusive",
+        "next generation",
+        "concept",
+        "approach"
+      ],
+      [
+        "disf",
+        "disintermediate",
+        "Grass-roots",
+        "policy",
+        "function"
+      ],
+      [
+        "cont",
+        "contingency",
+        "value-added",
+        "context-sensitive",
+        "Configurable"
+      ],
+      [
+        "Auto",
+        "human-resource",
+        "Reactive",
+        "hardware",
+        "Automated"
+      ],
+      [
+        "secon",
+        "Optional",
+        "secondary",
+        "object-oriented",
+        "toolset"
+      ]
+    ]
+
+	const key_value_1 = [
+	  [
+	    "static", # key
+	    "Inverse", # value
+	  ],
+	  [
+	    "function",
+	    "Horizontal",
+	  ],
+	  [
+	    "Integrated",
+	    "portal",
+	  ],
+	  [
+	    "non-based",
+	    "Monitored",
+	  ],
+	  [
+	    "productivity",
+	    "secondary",
+	  ],
+	  [
+	    "Exclusive",
+	    "next generation",
+	  ],
+	  [
+	    "disintermediate",
+	    "Grass-roots",
+	  ],
+	  [
+	    "contingency",
+	    "value-added",
+	  ],
+	  [
+	    "human-resource",
+	    "Reactive",
+	  ],
+	  [
+	    "Optional",
+	    "secondary",
+	  ]
+	]
+
+	const key_value_2 = [
+	  [
+	    "even-keeled", # key
+	    "Future-proofed" # value
+	  ],
+	  [
+	    "Face to face",
+	    "Expanded"
+	  ],
+	  [
+	    "budgetary management",
+	    "pricing structure"
+	  ],
+	  [
+	    "non-volatile",
+	    "emulation"
+	  ],
+	  [
+	    "Total",
+	    "Visionary"
+	  ],
+	  [
+	    "concept",
+	    "approach"
+	  ],
+	  [
+	    "policy",
+	    "function"
+	  ],
+	  [
+	    "context-sensitive",
+	    "Configurable"
+	  ],
+	  [
+	    "hardware",
+	    "Automated"
+	  ],
+	  [
+	    "object-oriented",
+	    "toolset"
+	  ]
+	]
+
+	const value_1 = [
+	  [
+	    "static", # value
+	  ],
+	  [
+	    "Inverse", # value...
+	  ],
+	  [
+	    "even-keeled",
+	  ],
+	  [
+	    "Future-proofed",
+	  ],
+	  [
+	    "function",
+	  ],
+	  [
+	    "Horizontal",
+	  ],
+	  [
+	    "Face to face",
+	  ],
+	  [
+	    "Expanded",
+	  ],
+	  [
+	    "Integrated",
+	  ],
+	  [
+	    "portal",
+	  ]
+	]
+
+    const value_2 = [
+      [
+        "even-keeled", # value
+      ],
+      [
+        "Face to face",
+      ],
+      [
+        "budgetary management",
+      ],
+      [
+        "non-volatile",
+      ],
+      [
+        "Total",
+      ],
+      [
+        "concept",
+      ],
+      [
+        "policy",
+      ],
+      [
+        "context-sensitive",
+      ],
+      [
+        "hardware",
+      ],
+      [
+        "object-oriented",
+      ]
+    ]
+
+	const offset = [
+	  [
+		3 # offset
+	  ],
+	  [
+		5
+	  ],
+	  [
+		0
+	  ],
+	  [
+		20
+	  ],
+	  [
+		1
+	  ],
+	  [
+		8
+	  ],
+	  [
+		6
+	  ],
+	  [
+		30
+	  ],
+	  [
+		2
+	  ],
+	  [
+		9
+	  ],
+	]
+
+	const child_child = [
+	  [
+	    ":child_storage:default:moratorium", # child1
+	    ":child_storage:default:hardware" # child2
+	  ],
+	  [
+	    ":child_storage:default:implementation",
+	    ":child_storage:default:artificial intelligence"
+	  ],
+	  [
+	    ":child_storage:default:Switchable",
+	    ":child_storage:default:matrix"
+	  ],
+	  [
+	    ":child_storage:default:strategy",
+	    ":child_storage:default:Graphic Interface"
+	  ],
+	  [
+	    ":child_storage:default:Total",
+	    ":child_storage:default:24 hour"
+	  ],
+	  [
+	    ":child_storage:default:exuding",
+	    ":child_storage:default:Triple-buffered"
+	  ],
+	  [
+	    ":child_storage:default:middleware",
+	    ":child_storage:default:Operative"
+	  ],
+	  [
+	    ":child_storage:default:Mandatory",
+	    ":child_storage:default:intranet"
+	  ],
+	  [
+	    ":child_storage:default:4th generation",
+	    ":child_storage:default:encryption"
+	  ],
+	  [
+	    ":child_storage:default:policy",
+	    ":child_storage:default:full-range"
+	  ]
+	]
+
+end
